Add aircraft component and mini views
